Move AppMiddleware to middleware module
